플레이어 공격턴 완성 및 버그 수정

+++ DiceKnight/Assets/Scripts/Data/DiceStats.cs
@@ -9,7 +9,7 @@ public class DiceStats : MonoBehaviour
 {
     public DiceType DiceType;
     public int Cost;
-    public int Hp;
+    public float Hp;
     public int Damage;
     public int Defense;
     public int Movement;
